feat: 예약이 존재하는 시간 삭제 시도 시, 예외 발생 구현
- 서비스 의존 방식에서 레포지토리 방식으로 변경
순환참조 문제 발생으로 인해 ApplicationService을 고려하였으나 새로운 계층 추가로 인한 복잡성이 더 커진다 생각하여 Repository 의존 방식으로 구현하였습니다.

+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -4,15 +4,18 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.domain.ReservationTime;
 import roomescape.dto.AddReservationTimeDto;
-import roomescape.exception.InvalidReservationTimeException;
+import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
 
 @Service
 public class ReservationTimeService {
 
+    private final ReservationRepository reservationRepository;
     private final ReservationTimeRepository reservationTimeRepository;
 
-    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository) {
+    public ReservationTimeService(ReservationRepository reservationRepository,
+                                  ReservationTimeRepository reservationTimeRepository) {
+        this.reservationRepository = reservationRepository;
         this.reservationTimeRepository = reservationTimeRepository;
     }
 
@@ -21,16 +24,14 @@ public class ReservationTimeService {
     }
 
     public void deleteReservationTime(Long id) {
+        if (reservationRepository.existsByTimeId(id)) {
+            throw new IllegalArgumentException("예약이 되어있는 시간은 삭제할 수 없습니다.");
+        }
         reservationTimeRepository.deleteById(id);
     }
 
     public Long addReservationTime(AddReservationTimeDto addReservationTimeDto) {
         ReservationTime reservationTime = addReservationTimeDto.toEntity();
         return reservationTimeRepository.add(reservationTime);
-    }
-
-    public ReservationTime findReservationTimeById(Long id) {
-        return reservationTimeRepository.findById(id)
-                .orElseThrow(() -> new InvalidReservationTimeException("존재하지 않는 id입니다."));
     }
 }
